Переделал отрисовку страницы Избранное (без карточек) по человечески
